Module refactored
