+++ platforms/windows/SQLitePlugin/ReactPackageProvider.cpp
@@ -6,7 +6,7 @@
 
 // NOTE: You must include the headers of your native modules here in
 // order for the AddAttributedModules call below to find them.
-#include "FancyMath.h"
+#include "SQLitePlugin.h"
 
 namespace winrt::SQLitePlugin::implementation
 {
