+++ platforms/windows/SQLitePlugin/SQLitePlugin.vcxproj.filters
@@ -12,10 +12,11 @@
   <ItemGroup>
     <ClCompile Include="pch.cpp" />
     <ClCompile Include="$(GeneratedFilesDir)module.g.cpp" />
+    <ClCompile Include="SQLitePlugin.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h" />
-    <ClInclude Include="FancyMath.h" />
+    <ClInclude Include="SQLitePlugin.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="SQLitePlugin.def" />
@@ -23,9 +24,6 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="PropertySheet.props" />
-  </ItemGroup>
-  <ItemGroup>
-    <Text Include="readme.txt" />
   </ItemGroup>
   <ItemGroup>
     <Midl Include="ReactPackageProvider.idl" />
