+++ platforms/windows/SQLitePlugin/pch.h
@@ -1,4 +1,12 @@
 ﻿#pragma once
 #include <unknwn.h>
+#include <winrt/Microsoft.ReactNative.h>
 #include <winrt/Windows.Foundation.h>
 #include <winrt/Windows.Foundation.Collections.h>
+#include <winrt/Windows.Security.Cryptography.h>
+#include <winrt/Windows.Storage.h>
+#include <winrt/Windows.Storage.Streams.h>
+#include <winsqlite/winsqlite3.h>
+
+#include "JSValue.h"
+#include "NativeModules.h"
